+++ examples/example-app-router-mixed-routing/src/app/(public)/[locale]/PublicNavigationLocaleSwitcher.tsx
@@ -2,7 +2,7 @@
 
 import {useLocale} from 'next-intl';
 import {Locale} from '@/config';
-import {Link, usePathname} from '@/routing.public';
+import {Link, usePathname} from '@/i18n/routing.public';
 
 export default function PublicNavigationLocaleSwitcher() {
   return (