+++ examples/example-app-router-mixed-routing/src/i18n/request.ts
@@ -1,15 +1,15 @@
 import {headers} from 'next/headers';
 import {notFound} from 'next/navigation';
 import {getRequestConfig} from 'next-intl/server';
-import {locales} from './config';
-import {getUserLocale} from './db';
+import {locales} from '../config';
+import {getUserLocale} from '../db';
 
 async function getConfig(locale: string) {
   // Validate that the incoming `locale` parameter is valid
   if (!locales.includes(locale as any)) notFound();
 
   return {
-    messages: (await import(`../messages/${locale}.json`)).default
+    messages: (await import(`../../messages/${locale}.json`)).default
   };
 }