+++ examples/example-app-router-mixed-routing/src/middleware.ts
@@ -1,6 +1,6 @@
 import {NextRequest, NextResponse} from 'next/server';
 import createMiddleware from 'next-intl/middleware';
-import {routing} from './routing.public';
+import {routing} from './i18n/routing.public';
 
 export default function middleware(request: NextRequest) {
   const pathname = request.nextUrl.pathname;