+++ examples/example-app-router-playground/src/components/LocaleSwitcher.tsx
@@ -1,5 +1,5 @@
 import {useLocale, useTranslations} from 'next-intl';
-import {Link} from '@/routing';
+import {Link} from '@/i18n/routing';
 
 export default function LocaleSwitcher() {
   const t = useTranslations('LocaleSwitcher');