+++ examples/example-app-router-playground/src/components/NavigationLink.tsx
@@ -2,7 +2,7 @@
 
 import {useSelectedLayoutSegment} from 'next/navigation';
 import {ComponentProps} from 'react';
-import {Link, Pathnames} from '@/routing';
+import {Link, Pathnames} from '@/i18n/routing';
 
 export default function NavigationLink<Pathname extends Pathnames>({
   href,