+++ examples/example-app-router/src/components/LocaleSwitcherSelect.tsx
@@ -3,7 +3,7 @@
 import clsx from 'clsx';
 import {useParams} from 'next/navigation';
 import {ChangeEvent, ReactNode, useTransition} from 'react';
-import {Locale, usePathname, useRouter} from '@/routing';
+import {Locale, usePathname, useRouter} from '@/i18n/routing';
 
 type Props = {
   children: ReactNode;